+++ packages/contact-center/cc-components/src/components/task/CallControl/call-control.tsx
@@ -1,19 +1,18 @@
 import React, {useEffect, useMemo, useState} from 'react';
 
-import {CallControlComponentProps} from '../task.types';
+import {CallControlComponentProps, DestinationType, CallControlMenuType} from '../task.types';
 import './call-control.styles.scss';
 import {PopoverNext, SelectNext, TooltipNext, Text, ButtonCircle, ButtonPill} from '@momentum-ui/react-collaboration';
 import {Item} from '@react-stately/collections';
 import {Icon} from '@momentum-design/components/dist/react';
 import ConsultTransferPopoverComponent from './CallControlCustom/consult-transfer-popover';
-import CallControlConsultComponent from './CallControlCustom/call-control-consult';
 import {getControlsVisibility} from '../../../utils/task-util';
 
 function CallControlComponent(props: CallControlComponentProps) {
   const [selectedWrapupReason, setSelectedWrapupReason] = useState<string | null>(null);
   const [selectedWrapupId, setSelectedWrapupId] = useState<string | null>(null);
   const [showAgentMenu, setShowAgentMenu] = useState(false);
-  const [agentMenuType, setAgentMenuType] = useState<'Consult' | 'Transfer' | null>(null);
+  const [agentMenuType, setAgentMenuType] = useState<CallControlMenuType | null>(null);
 
   const {
     currentTask,
@@ -29,21 +28,19 @@ function CallControlComponent(props: CallControlComponentProps) {
     setIsRecording,
     buddyAgents,
     loadBuddyAgents,
+    queues,
+    loadQueues,
     transferCall,
     consultCall,
-    endConsultCall,
-    consultTransfer,
     consultInitiated,
-    consultCompleted,
     consultAccepted,
-    consultStartTimeStamp,
     callControlAudio,
-    consultAgentName,
     setConsultAgentName,
-    consultAgentId,
     setConsultAgentId,
     deviceType,
     featureFlags,
+    allowConsultToQueue,
+    setLastTargetType,
   } = props;
 
   const visibleCtrl = useMemo(
@@ -83,13 +80,34 @@ function CallControlComponent(props: CallControlComponentProps) {
     setSelectedWrapupId(value);
   };
 
-  const handleAgentSelect = (agentId: string, agentName: string) => {
+  const handleTargetSelect = (id: string, name: string, type: DestinationType) => {
     if (agentMenuType === 'Consult') {
-      consultCall(agentId, 'agent');
-      setConsultAgentId(agentId);
-      setConsultAgentName(agentName);
+      try {
+        consultCall(id, type);
+        setConsultAgentId(id);
+        setConsultAgentName(name);
+        setLastTargetType(type);
+      } catch (error) {
+        throw new Error('Error during consult call', error);
+      }
+    } else if (agentMenuType === 'Transfer') {
+      try {
+        transferCall(id, type);
+      } catch (error) {
+        throw new Error('Error during transfer call', error);
+      }
+    }
+  };
+
+  const handlePopoverOpen = (menuType: CallControlMenuType) => {
+    if (showAgentMenu && agentMenuType === menuType) {
+      setShowAgentMenu(false);
+      setAgentMenuType(null);
     } else {
-      transferCall(agentId, 'agent');
+      setAgentMenuType(menuType);
+      setShowAgentMenu(true);
+      loadBuddyAgents();
+      loadQueues();
     }
   };
 
@@ -199,16 +217,7 @@ function CallControlComponent(props: CallControlComponentProps) {
                             aria-label={button.tooltip}
                             disabled={button.disabled || consultInitiated}
                             data-testid="ButtonCircle"
-                            onPress={() => {
-                              if (showAgentMenu && agentMenuType === button.menuType) {
-                                setShowAgentMenu(false);
-                                setAgentMenuType(null);
-                              } else {
-                                setAgentMenuType(button.menuType as 'Consult' | 'Transfer');
-                                setShowAgentMenu(true);
-                                loadBuddyAgents();
-                              }
-                            }}
+                            onPress={() => handlePopoverOpen(button.menuType as CallControlMenuType)}
                           >
                             <Icon className={button.className + '-icon'} name={button.icon} />
                           </ButtonCircle>
@@ -229,7 +238,10 @@ function CallControlComponent(props: CallControlComponentProps) {
                         heading={button.menuType}
                         buttonIcon={button.icon}
                         buddyAgents={buddyAgents}
-                        onAgentSelect={handleAgentSelect}
+                        queues={queues}
+                        onAgentSelect={(agentId, agentName) => handleTargetSelect(agentId, agentName, 'agent')}
+                        onQueueSelect={(queueId, queueName) => handleTargetSelect(queueId, queueName, 'queue')}
+                        allowConsultToQueue={allowConsultToQueue}
                       />
                     ) : null}
                   </PopoverNext>
@@ -315,19 +327,6 @@ function CallControlComponent(props: CallControlComponentProps) {
                 Submit & Wrap up
               </ButtonPill>
             </PopoverNext>
-          </div>
-        )}
-
-        {(consultAccepted || consultInitiated) && !wrapupRequired && (
-          <div className={`call-control-consult-container ${consultAccepted ? 'no-border' : ''}`}>
-            <CallControlConsultComponent
-              agentName={consultAgentName}
-              startTimeStamp={consultStartTimeStamp}
-              endConsultCall={endConsultCall}
-              onTransfer={() => consultTransfer(consultAgentId || currentTask.data.destAgentId, 'agent')}
-              consultCompleted={consultCompleted}
-              showTransfer={!consultAccepted}
-            />
           </div>
         )}
       </div>