+++ packages/contact-center/station-login/src/station-login/index.tsx
@@ -6,23 +6,39 @@ import {StationLoginComponent} from '@webex/cc-components';
 import {useStationLogin} from '../helper';
 import {StationLoginProps} from './station-login.types';
 
-const StationLogin: React.FunctionComponent<StationLoginProps> = observer(({onLogin, onLogout}) => {
-  const {cc, teams, loginOptions, logger, isAgentLoggedIn, showMultipleLoginAlert, deviceType} = store;
+const StationLogin: React.FunctionComponent<StationLoginProps> = observer(({onLogin, onLogout, onCCSignOut}) => {
+  const {
+    cc,
+    teams,
+    loginOptions,
+    logger,
+    isAgentLoggedIn,
+    showMultipleLoginAlert,
+    deviceType,
+    dialNumber,
+    setDeviceType,
+    setDialNumber,
+  } = store;
   const result = useStationLogin({
     cc,
     onLogin,
     onLogout,
     logger,
     deviceType,
+    dialNumber,
   });
 
   const props = {
     ...result,
+    setDeviceType,
+    setDialNumber,
     teams,
     loginOptions,
     deviceType,
+    dialNumber,
     isAgentLoggedIn,
     showMultipleLoginAlert,
+    onCCSignOut,
   };
   return <StationLoginComponent {...props} />;
 });