+++ packages/contact-center/store/src/store.ts
@@ -30,6 +30,7 @@ class Store implements IStore {
   currentTask: ITask = null;
   isAgentLoggedIn = false;
   deviceType: string = '';
+  dialNumber: string = '';
   taskList: ITask[] = [];
   wrapupRequired: boolean = false;
   currentState: string = '';
@@ -37,6 +38,8 @@ class Store implements IStore {
   consultCompleted = false;
   consultInitiated = false;
   consultAccepted = false;
+  isQueueConsultInProgress = false;
+  currentConsultQueueId: string = '';
   consultStartTimeStamp = undefined;
   lastStateChangeTimestamp?: number;
   lastIdleCodeChangeTimestamp?: number;
@@ -44,6 +47,8 @@ class Store implements IStore {
   callControlAudio: MediaStream | null = null;
   consultOfferReceived: boolean = false;
   featureFlags: {[key: string]: boolean} = {};
+  isEndConsultEnabled: boolean = false;
+  allowConsultToQueue: boolean = false;
 
   constructor() {
     makeAutoObservable(this, {
@@ -82,10 +87,13 @@ class Store implements IStore {
         this.agentId = response.agentId;
         this.wrapupCodes = response.wrapupCodes;
         this.isAgentLoggedIn = response.isAgentLoggedIn;
-        this.deviceType = response.deviceType;
+        this.deviceType = response.deviceType ?? 'AGENT_DN';
+        this.dialNumber = response.defaultDn;
         this.currentState = response.lastStateAuxCodeId;
         this.lastStateChangeTimestamp = response.lastStateChangeTimestamp;
         this.lastIdleCodeChangeTimestamp = response.lastIdleCodeChangeTimestamp;
+        this.isEndConsultEnabled = response.isEndConsultEnabled;
+        this.allowConsultToQueue = response.allowConsultToQueue;
       })
       .catch((error) => {
         this.logger.error(`Error registering contact center: ${error}`, {