+++ packages/contact-center/store/src/storeEventsWrapper.ts
@@ -13,6 +13,8 @@ import {
   BuddyDetails,
   ENGAGED_LABEL,
   ENGAGED_USERNAME,
+  ContactServiceQueue,
+  Profile,
 } from './store.types';
 import Store from './store';
 import {runInAction} from 'mobx';
@@ -53,6 +55,9 @@ class StoreWrapper implements IStoreWrapper {
 
   get deviceType() {
     return this.store.deviceType;
+  }
+  get dialNumber() {
+    return this.store.dialNumber;
   }
   get wrapupCodes() {
     return this.store.wrapupCodes;
@@ -123,6 +128,22 @@ class StoreWrapper implements IStoreWrapper {
     return this.store.consultOfferReceived;
   }
 
+  get isQueueConsultInProgress() {
+    return this.store.isQueueConsultInProgress;
+  }
+
+  get currentConsultQueueId() {
+    return this.store.currentConsultQueueId;
+  }
+
+  get isEndConsultEnabled() {
+    return this.store.isEndConsultEnabled;
+  }
+
+  get allowConsultToQueue() {
+    return this.store.allowConsultToQueue;
+  }
+
   setCurrentTheme = (theme: string): void => {
     this.store.currentTheme = theme;
   };
@@ -133,6 +154,10 @@ class StoreWrapper implements IStoreWrapper {
 
   setDeviceType = (option: string): void => {
     this.store.deviceType = option;
+  };
+
+  setDialNumber = (input: string): void => {
+    this.store.dialNumber = input;
   };
 
   setCurrentState = (state: string): void => {
@@ -204,6 +229,18 @@ class StoreWrapper implements IStoreWrapper {
     this.store.callControlAudio = audio;
   };
 
+  setIsQueueConsultInProgress = (value: boolean): void => {
+    runInAction(() => {
+      this.store.isQueueConsultInProgress = value;
+    });
+  };
+
+  setCurrentConsultQueueId = (queueId: string | null): void => {
+    runInAction(() => {
+      this.store.currentConsultQueueId = queueId;
+    });
+  };
+
   setState = (state: ICustomState | IdleCode): void => {
     if ('reset' in state) {
       runInAction(() => {
@@ -264,7 +301,7 @@ class StoreWrapper implements IStoreWrapper {
     const taskToRemove = this.store.taskList.find((task) => task.data.interactionId === taskId);
     if (taskToRemove) {
       taskToRemove.off(TASK_EVENTS.TASK_ASSIGNED, this.handleTaskAssigned);
-      taskToRemove.off(TASK_EVENTS.TASK_END, ({wrapupRequired}) => this.handleTaskEnd(taskToRemove, wrapupRequired));
+      taskToRemove.off(TASK_EVENTS.TASK_END, this.handleTaskEnd);
       taskToRemove.off(TASK_EVENTS.TASK_REJECT, (reason) => this.handleTaskReject(taskToRemove.interactionId, reason));
       taskToRemove.off(TASK_EVENTS.AGENT_WRAPPEDUP, this.handleTaskWrapUp);
       taskToRemove.off(TASK_EVENTS.TASK_CONSULTING, this.handleConsulting);
@@ -272,6 +309,7 @@ class StoreWrapper implements IStoreWrapper {
       taskToRemove.off(TASK_EVENTS.TASK_CONSULT_END, this.handleConsultEnd);
       taskToRemove.off(TASK_EVENTS.TASK_CONSULT_ACCEPTED, this.handleConsultAccepted);
       taskToRemove.off(TASK_EVENTS.AGENT_CONSULT_CREATED, this.handleConsultCreated);
+      taskToRemove.off(TASK_EVENTS.TASK_CONSULT_QUEUE_CANCELLED, this.handleConsultQueueCancelled);
       if (this.deviceType === 'BROWSER') {
         taskToRemove.off(TASK_EVENTS.TASK_MEDIA, this.handleTaskMedia);
         this.setCallControlAudio(null);
@@ -302,12 +340,12 @@ class StoreWrapper implements IStoreWrapper {
     });
   };
 
-  handleTaskEnd = (event, wrapupRequired) => {
+  handleTaskEnd = (event) => {
     // If the call is ended by agent we get the task object in event.data
     // If the call is ended by customer we get the task object directly
 
     const task = event.data ? event.data : event;
-    if (wrapupRequired) {
+    if (task.wrapUpRequired) {
       this.setWrapupRequired(true);
     } else {
       this.handleTaskRemove(task.interactionId);
@@ -321,6 +359,7 @@ class StoreWrapper implements IStoreWrapper {
         this.setConsultAccepted(false);
         this.setConsultInitiated(false);
         this.setConsultCompleted(false);
+        this.setConsultOfferReceived(false);
       }
       this.setCurrentTask(task);
       this.setIncomingTask(null);
@@ -357,6 +396,8 @@ class StoreWrapper implements IStoreWrapper {
   handleConsultEnd = (event) => {
     const task = event;
     this.setConsultInitiated(false);
+    this.setIsQueueConsultInProgress(false);
+    this.setCurrentConsultQueueId(null);
     if (this.consultAccepted) {
       this.setConsultAccepted(false);
       this.handleTaskRemove(task.data.interactionId);
@@ -390,6 +431,13 @@ class StoreWrapper implements IStoreWrapper {
     });
   };
 
+  handleConsultQueueCancelled = () => {
+    this.setConsultInitiated(false);
+    this.setIsQueueConsultInProgress(false);
+    this.setCurrentConsultQueueId(null);
+    this.setConsultStartTimeStamp(null);
+  };
+
   handleIncomingTask = (event) => {
     const task: ITask = event;
     if (this.store.taskList.some((t) => t.data.interactionId === task.data.interactionId)) {
@@ -398,11 +446,12 @@ class StoreWrapper implements IStoreWrapper {
     }
 
     // Attach event listeners to the task
-    task.on(TASK_EVENTS.TASK_END, ({wrapupRequired}) => this.handleTaskEnd(task, wrapupRequired));
+    task.on(TASK_EVENTS.TASK_END, () => this.handleTaskEnd(task));
 
     // When we receive TASK_ASSIGNED the task was accepted by the agent and we need wrap up
     task.on(TASK_EVENTS.TASK_ASSIGNED, this.handleTaskAssigned);
     task.on(TASK_EVENTS.AGENT_CONSULT_CREATED, this.handleConsultCreated);
+    task.on(TASK_EVENTS.TASK_CONSULT_QUEUE_CANCELLED, this.handleConsultQueueCancelled);
     if (this.deviceType === 'BROWSER') {
       task.on(TASK_EVENTS.TASK_MEDIA, this.handleTaskMedia);
     }
@@ -440,7 +489,7 @@ class StoreWrapper implements IStoreWrapper {
 
   handleTaskHydrate = (event) => {
     const task = event;
-    task.on(TASK_EVENTS.TASK_END, ({wrapupRequired}) => this.handleTaskEnd(task, wrapupRequired));
+    task.on(TASK_EVENTS.TASK_END, () => this.handleTaskEnd(task));
 
     // When we receive TASK_ASSIGNED the task was accepted by the agent and we need wrap up
     task.on(TASK_EVENTS.TASK_ASSIGNED, this.handleTaskAssigned);
@@ -456,6 +505,7 @@ class StoreWrapper implements IStoreWrapper {
     task.on(TASK_EVENTS.TASK_CONSULTING, this.handleConsulting);
     task.on(CC_EVENTS.AGENT_OFFER_CONSULT, this.handleConsultOffer);
     task.on(TASK_EVENTS.TASK_CONSULT_END, this.handleConsultEnd);
+    task.on(TASK_EVENTS.TASK_CONSULT_QUEUE_CANCELLED, this.handleConsultQueueCancelled);
     if (this.deviceType === 'BROWSER') {
       task.on(TASK_EVENTS.TASK_MEDIA, this.handleTaskMedia);
     }
@@ -481,16 +531,15 @@ class StoreWrapper implements IStoreWrapper {
       name: ENGAGED_USERNAME,
     });
 
-    const {interaction, agentId} = task.data;
-    const {state, isTerminated, participants} = interaction;
+    const {interaction} = task.data;
+    const {isTerminated} = interaction;
 
     // Update call control states
     if (isTerminated) {
       // wrapup
-      const wrapupRequired = state === 'wrapUp' && !participants[agentId].isWrappedUp;
-      this.setWrapupRequired(wrapupRequired);
+      this.setWrapupRequired(task.data.wrapUpRequired);
 
-      if (!wrapupRequired) {
+      if (!task.data.wrapUpRequired) {
         this.setState({
           reset: true,
         });
@@ -519,10 +568,22 @@ class StoreWrapper implements IStoreWrapper {
     }
   };
 
+  getQueues = async (): Promise<Array<ContactServiceQueue>> => {
+    try {
+      let queueList = await this.store.cc.getQueues();
+      queueList = queueList.filter((queue) => queue.channelType === 'TELEPHONY');
+      return queueList;
+    } catch (error) {
+      console.error('Error fetching queues:', error);
+      return Promise.reject(error);
+    }
+  };
+
   cleanUpStore = () => {
     runInAction(() => {
       this.setIsAgentLoggedIn(false);
-      this.setDeviceType('');
+      this.setDeviceType('AGENT_DN');
+      this.setDialNumber('');
       this.setIncomingTask(null);
       this.setCurrentTask(null);
       this.setTaskList([]);
@@ -561,10 +622,11 @@ class StoreWrapper implements IStoreWrapper {
 
     // TODO: https://jira-eng-gpk2.cisco.com/jira/browse/SPARK-626777 Implement the de-register method and close the listener there
 
-    const handleLogin = (payload) => {
+    const handleLogin = (payload: Profile) => {
       runInAction(() => {
         this.setIsAgentLoggedIn(true);
         this.setDeviceType(payload.deviceType);
+        this.setDialNumber(payload.dn);
         this.setCurrentState(payload.auxCodeId?.trim() !== '' ? payload.auxCodeId : '0');
         this.setLastStateChangeTimestamp(payload.lastStateChangeTimestamp);
         this.setLastIdleCodeChangeTimestamp(payload.lastIdleCodeChangeTimestamp);