+++ packages/contact-center/task/tests/TaskList/index.tsx
@@ -16,6 +16,7 @@ jest.mock('@webex/cc-components', () => {
 jest.mock('@webex/cc-store', () => ({
   cc: {},
   deviceType: 'BROWSER',
+  dialNumber: '12345',
   onAccepted: jest.fn(),
   onDeclined: jest.fn(),
 }));